spoofing_detection_api/requirements.txt
fastapi==0.128.0
numpy==2.4.1
Pillow==12.1.0
pydantic==2.12.5
pydantic_settings==2.12.0
spoofing_recognition==0.1.0
spoofing_recognition.egg==info
torch==2.9.1

-e /code/app
